// api/src/types/CreateOrganizationBulkDTO.ts
export interface CreateOrganizationBulkDTO {
    name: string;
    location1?: string;
    address1?: string;
    location2?: string;
    address2?: string;
    location3?: string;
    address3?: string;
    location4?: string;
    address4?: string;
    location5?: string;
    address5?: string;
    aliases: string[];
  }